+++ src/modules/Preservation/views/AddScope/SelectTags/SelectTags.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 
 import { Button, TextField } from '@equinor/eds-core-react';
 import CheckBoxIcon from '@material-ui/icons/CheckBox';
-import { Tag, TagRow } from './types';
+import { Tag, TagRow } from '../types';
 import { Container, Header, Actions, Search, Next, Tags, TagsHeader, LoadingContainer, Toolbar } from './SelectTags.style';
-import { usePreservationContext } from '../../context/PreservationContext';
-import Table from './../../../../components/Table';
-import Loading from './../../../../components/Loading';
+import { usePreservationContext } from '../../../context/PreservationContext';
+import Table from '../../../../../components/Table';
+import Loading from '../../../../../components/Loading';
 
 type SelectTagsProps = {
     selectedTags: Tag[];
